Move broker_conformance_test back to e2e_new
